[FIX] nicknameState분리(#17)
